Do not run a test in parallel with others
Do not run the test for DeleteUserInstances() in parallel with the other tests.
